Fix unsigned/signed int comparison compiler warning
Regarding the variable type changes in kernel_image, I don't think we'll
need image dimensions that need a size_t to fit in the foreseeable
future. Same for the image_levels_.

+++ se_voxel_impl/include/se/voxel_implementations/MultiresOFusion/kernel_image.hpp
@@ -122,16 +122,16 @@ namespace se {
     KernelImage::Pixel conservativeQuery(const Eigen::Vector2i& bb_min, const Eigen::Vector2i& bb_max) const;
     KernelImage::Pixel poolBoundingBox(int u_min, int u_max, int v_min, int v_max) const;
 
-    size_t width() const {return image_width_;};
-    size_t height() const {return image_height_;};
-    Value  maxValue() const {return image_max_value_;};
-    size_t maxLevel() const {return image_max_level_;}
+    int width() const {return image_width_;};
+    int height() const {return image_height_;};
+    Value maxValue() const {return image_max_value_;};
+    int maxLevel() const {return image_max_level_;}
 
   private:
-    size_t    image_max_level_;
-    size_t    image_width_;
-    size_t    image_height_;
-    Pyramid  pyramid_image_;
+    int     image_max_level_;
+    int     image_width_;
+    int     image_height_;
+    Pyramid pyramid_image_;
     Value   image_max_value_;
   };
